+++ R/varstan.R
@@ -102,12 +102,12 @@
 #'  library(astsa)
 #'  # Fitting a seasonal ARIMA model
 #'  mod1 = Sarima(birth,order = c(0,1,2),seasonal = c(1,1,1))
-#'  fit1 = varstan(mod1,chains = 1)
+#'  fit1 = varstan(mod1,iter = 500,chains = 1)
 #'  fit1
 #'
 #'  # Fitting a GARCH(1,1) model
 #'  dat = garch(ipc,order = c(1,1,0))
-#'  fit2 = varstan(dat,chains = 1)
+#'  fit2 = varstan(dat,iter = 500,chains = 1)
 #'  fit2
 #' }
 #'
@@ -124,6 +124,9 @@ varstan = function(model,chains = 4,iter = 2000,warmup = floor(iter/2),
   if(is.garch(model))  sft = fit_garch(model,chains,iter,warmup,adapt.delta,tree.depth)
   if(is.SVM(model))    sft = fit_SVM(model,chains,iter,warmup,adapt.delta,tree.depth)
   if(is.ssm(model))    sft = fit_ssm(model,chains,iter,warmup,adapt.delta,tree.depth)
+  if(is.LocalLevel(model))sft = fit_ssm(model,chains,iter,warmup,adapt.delta,tree.depth)
+  if(is.Holt(model))   sft = fit_ssm(model,chains,iter,warmup,adapt.delta,tree.depth)
+  if(is.Hw(model))     sft = fit_ssm(model,chains,iter,warmup,adapt.delta,tree.depth)
 
   sp = list(Algorithm = "HMC NUTS",chains = chains,iter = iter,warmup = warmup,
             adapt.delta =adapt.delta,max_treedepth = tree.depth)
@@ -184,7 +187,7 @@ is.varstan = function(object){
 #'  library(astsa)
 #'  # Fitting a GARCH(1,1) model
 #'  dat = garch(ipc,order = c(1,1,0))
-#'  fit2 = varstan(dat,chains = 1)
+#'  fit2 = varstan(dat,iter = 500,chains = 1)
 #'
 #'  # Extracting the mean parameter
 #'  mu0 = extract_stan(fit2,pars = "mu0")
@@ -218,7 +221,7 @@ extract_stan = function(object,pars,permuted = TRUE,
 #' \donttest{
 #'  # Fitting a GARCH(1,1) model
 #'  dat = garch(ipc,order = c(1,1,0))
-#'  fit1 = varstan(dat,chains = 1)
+#'  fit1 = varstan(dat,iter = 500,chains = 1)
 #'
 #'  # Converting to a Stanfit object
 #'  stanfit1 = as.stan(fit1)